requirements.txt
mcp>=1.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-compress>=1.14
gunicorn>=21.0.0
requests>=2.31.0
httpx>=0.27.0
feedparser>=6.0.0
python-dateutil>=2.9.0
